feat(client): modal component #13
modal component 1차 수정했습니다 !
