+++ apps/web/src/components/common/Modal/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Modal';
+export * from './types';
