+++ apps/web/src/components/common/Modal/variants/index.ts
@@ -1,0 +1,4 @@
+export * from './MessageModal';
+export * from './ConfirmModal';
+export * from './InputModal';
+export * from './ProfileModal';
