+++ package.json
@@ -31,6 +31,7 @@
     "node": ">=20"
   },
   "dependencies": {
+    "ts-pattern": "^5.6.2",
     "zod": "^3.24.1"
   }
 }
